+++ src/AWS/Lambda/Context.hs
@@ -15,21 +15,15 @@ module AWS.Lambda.Context (
   CognitoIdentity(..),
   LambdaContext(..),
   getRemainingTime,
-  HasLambdaContext(..),
-  defConfig,
-  runReaderTLambdaContext
 ) where
 
 import           Control.Monad.IO.Class (MonadIO, liftIO)
-import           Control.Monad.Reader   (ReaderT, runReaderT)
 import           Data.Aeson             (FromJSON, ToJSON)
 import           Data.Map               (Map)
 import           Data.Text              (Text)
 import           Data.Time.Clock        (DiffTime, UTCTime,
                                          diffUTCTime, getCurrentTime)
-import           Data.Time.Clock.POSIX  (posixSecondsToUTCTime)
 import           GHC.Generics           (Generic)
-import           System.Envy            (DefConfig (..))
 
 data ClientApplication = ClientApplication
   { appTitle       :: Text,
@@ -76,23 +70,3 @@ data LambdaContext = LambdaContext
     clientContext      :: Maybe ClientContext,
     identity           :: Maybe CognitoIdentity
   } deriving (Show, Generic, Eq)
-
-{-# DEPRECATED HasLambdaContext "HasLambdaContext will be removed along with the original mRuntimeWithContext.  This utility is no longer necessary without it." #-}
-class HasLambdaContext r where
-  withContext :: (LambdaContext -> r -> r)
-
-instance HasLambdaContext LambdaContext where
-  withContext = const
-
--- TODO: This sticks around for backwards compatibility, and as a conevient-ish
--- way to runReaderTLambdaContext.  In the long term, all runtimes where the
--- LambdaContext is (incorrectly) available outside of the request/response
--- cycle will be removed.  This instance (and its dependent package, envy) can
--- be dropped on that breaking change.
-instance DefConfig LambdaContext where
-  defConfig = LambdaContext "" "" 0 "" "" "" "" "" (posixSecondsToUTCTime 0) Nothing Nothing
-
--- | Helper for using arbitrary monads with only the LambdaContext in its Reader
-{-# DEPRECATED runReaderTLambdaContext "runReaderTLambdaContext will be removed along with the original mRuntimeWithContext.  This particular approach was problematic, in that it required a default LambdaContext, when in reality, there is no valid instance." #-}
-runReaderTLambdaContext :: ReaderT LambdaContext m a -> m a
-runReaderTLambdaContext = flip runReaderT defConfig